Abstract `onChangeLayer` in order to use it to update DC
