+++ src/components/background-control/backgroundLayersControl.js
@@ -23,7 +23,7 @@ import Icon from '../icon';
 import ImageLayer from './imageLayer';
 import SVGLayer from './svgLayer';
 import VideoLayer from './videoLayer';
-import { setBreakpointToLayer } from './utils';
+import { onChangeLayer, setBreakpointToLayer } from './utils';
 import SelectControl from '../select-control';
 import ListControl from '../list-control';
 import ListItemControl from '../list-control/list-item-control';
@@ -32,15 +32,7 @@ import ListItemControl from '../list-control/list-item-control';
  * External dependencies
  */
 import classnames from 'classnames';
-import {
-	cloneDeep,
-	findIndex,
-	isEmpty,
-	isEqual,
-	isNil,
-	omit,
-	omitBy,
-} from 'lodash';
+import { cloneDeep, isEmpty, omit } from 'lodash';
 
 /**
  * Icons
@@ -485,29 +477,6 @@ const BackgroundLayersControl = ({
 			},
 		});
 
-	const onChangeLayer = (rawLayer, target = false) => {
-		const layer = omitBy(rawLayer, isNil);
-		const isHoverLayer = layer.isHover;
-		const newLayers = cloneDeep(isHoverLayer ? layersHover : layers);
-
-		allLayers.forEach((lay, i) => {
-			if (lay.order === layer.order) {
-				const index = findIndex(newLayers, { order: layer.order });
-
-				newLayers[index] = layer;
-			}
-		});
-
-		if (!isEqual(newLayers, isHoverLayer ? layersHover : layers))
-			onChange(
-				{
-					[`background-layers${isHoverLayer ? '-hover' : ''}`]:
-						newLayers,
-				},
-				target
-			);
-	};
-
 	const onAddLayer = layer => {
 		const isHoverLayer = layer.isHover;
 		const newLayers = cloneDeep(isHoverLayer ? layersHover : layers);
@@ -569,7 +538,14 @@ const BackgroundLayersControl = ({
 										isHover,
 										clientId,
 										layer,
-										onChange: onChangeLayer,
+										onChange: (rawLayer, target = false) =>
+											onChangeLayer(
+												rawLayer,
+												onChange,
+												layersHover,
+												layers,
+												target
+											),
 										breakpoint,
 										handleOnChangeLayer,
 										previewRef,
@@ -581,7 +557,14 @@ const BackgroundLayersControl = ({
 										isIB,
 										layer,
 										onChangeInline,
-										onChange: onChangeLayer,
+										onChange: (rawLayer, target = false) =>
+											onChangeLayer(
+												rawLayer,
+												onChange,
+												layersHover,
+												layers,
+												target
+											),
 										previewRef,
 										getBounds,
 										getBlockClipPath, // for IB
